backend/requirements.txt
openai==1.30.1
opencv-python
requests
pillow
sewar
screeninfo
pyperclip
scikit-image
langdetect
fasttext-langdetect
profanity-check
alt-profanity-check
scikit-learn
joblib
fastapi
uvicorn[standard]
sqlalchemy
databases
passlib[bcrypt]
pydantic[email]
python-jose[cryptography]
python-multipart
factory_boy